Add identity client distributed tracing instrumentation
